refactor(server): dedupe rank-page-row mirror types and optional-override spread
rls/middleware.ts and mask/middleware.ts each hand-maintained a byte-identical
copy of the rankPageRows result shape (RankPageRowKeyLike/RankPageRowLike/
ShardRankPageResultLike) — two mirrors of one upstream shard-engine type that
can drift apart unnoticed. Both now import a single shared module instead.

Also collapses the `...(base ? { [key]: ... } : {})` incantation, written by
hand four times across rankBefore/rankPageRows in both middlewares, into one
optionalWriterOverride helper. This shrinks mask/middleware.ts back under the
repo's file-size convention without changing behavior.
